pyproject.toml
[tool.poetry]
name = "hive"
version = "0.1.0"
description = ""
authors = ["de1ux <nathan@de1ux.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
Django = "^3.2.9"
python-env-utils = "^0.4.1"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.2"
reportlab = "^3.6.2"
qrcode = "^7.3.1"
django-cors-headers = "^3.7.0"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
fourmat = "^0.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
